fix: resolve three golangci-lint issues flagged by CI
- G115 (gosec): blocksLeft atomic.Int32 → Int64 eliminates int→int32 cast
- G115 (gosec): os.Getpid() routed through uint before uint64 cast
- SA1019 (staticcheck): replace deprecated parser.ParseDir with
  parser.ParseFile + doc.NewFromFiles in schedule doc_test.go

+++ cmd/phantom/run.go
@@ -108,7 +108,7 @@ func runRun(args []string, stdout, stderr io.Writer) int {
 	// the forbidigo time.Now boundary is not crossed inside the package.
 	r := rng.New(
 		uint64(time.Now().UnixNano()),
-		uint64(os.Getpid())*0x9e3779b97f4a7c15,
+		uint64(uint(os.Getpid()))*0x9e3779b97f4a7c15,
 	)
 
 	e, err := engine.New(engine.Options{
